requirements.txt
email-validator==2.3.0
fastapi==0.120.1
fastapi-pagination==0.14.3
pydantic==2.12.3
pytest==8.4.2
python-dotenv==1.2.1
requests==2.32.5
uvicorn==0.38.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.44
sqlmodel==0.0.27
faker==37.12.0